Remove all SnapshotOp integrations
