--- hedera-node/test-clients/src/main/java/com/hedera/services/bdd/spec/utilops/records/SnapshotMode.java
@@ -1,43 +1,0 @@
-/*
- * Copyright (C) 2023-2024 Hedera Hashgraph, LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.hedera.services.bdd.spec.utilops.records;
-
-import com.hedera.services.bdd.spec.HapiSpec;
-import com.hedera.services.bdd.spec.TargetNetworkType;
-
-public enum SnapshotMode {
-    /**
-     * Takes a snapshot of the record stream generated by running a {@link HapiSpec} against a
-     * {@link com.hedera.services.bdd.junit.HapiTest} network.
-     */
-    TAKE_FROM_HAPI_TEST_STREAMS(TargetNetworkType.SUBPROCESS_NETWORK),
-    /**
-     * Fuzzy-matches the record stream generated by running a {@link HapiSpec} against a
-     * {@link com.hedera.services.bdd.junit.HapiTest} network with a saved snapshot.
-     */
-    FUZZY_MATCH_AGAINST_HAPI_TEST_STREAMS(TargetNetworkType.SUBPROCESS_NETWORK);
-
-    private final TargetNetworkType targetNetworkType;
-
-    SnapshotMode(TargetNetworkType targetNetworkType) {
-        this.targetNetworkType = targetNetworkType;
-    }
-
-    public TargetNetworkType targetNetworkType() {
-        return targetNetworkType;
-    }
-}
